+++ src/weak_ctx_.d.ts
@@ -1,1 +1,0 @@
-export declare function weak_ctx_<Ctx extends object = object>(back_ctx?: {}): Ctx;